requirements.txt
# GCP
google-cloud-datastore
appengine-python-standard
google-cloud-logging
google_api_python_client

# Web
beautifulsoup4
gunicorn
flask
jinja2
pyjwt
webapp2

# Testing
WebTest
mock

# Libraries
oauth2client
rsa
pycrypto

# Development
pbr
black
isort